refactor: move runPerformanceCli into cli lib
