+++ src/cli/commands/bench-identifier-pipeline.mjs
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 import process from "node:process";
 
-import { runPerformanceCli } from "../src/cli/performance.js";
+import { runPerformanceCli } from "../lib/performance-cli.js";
 
 const argv = process.argv.slice(2);
