+++ src/cli/commands/measure-project-index-memory.mjs
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 import process from "node:process";
 
-import { runPerformanceCli } from "../src/cli/performance.js";
+import { runPerformanceCli } from "../lib/performance-cli.js";
 
 const exitCode = await runPerformanceCli({
     argv: ["--suite", "project-index-memory", ...process.argv.slice(2)]
